requirements.txt
requests
python-dotenv
firecrawl-py
openai
